backend/requirements.txt
Django==2.0.6
django-mysql==2.4.1
django-cors-headers==2.4.0
django-multiselectfield==0.1.8
djangorestframework==3.8.2
requests
mysqlclient==1.3.13
python-decouple==3.1

